requirements.txt
flask
flask-cors
openai
streamlit
scikit-learn
spacy
nltk
python-dotenv
requests